test_api/requirements.txt
django==3.2.8
djangorestframework==3.12.4
drf-yasg==1.20.0
PyJWT==2.1.0
python-dotenv==0.19.2
psycopg2-binary==2.9.3
gunicorn==20.0.4
djoser==2.1.0 
flake8==4.0.1


